Adjustments to make cog easier to understand.
